Create dibo package
