+++ src/main/java/dibo/ByeCommand.java
@@ -1,3 +1,4 @@
+package dibo;
 public class ByeCommand extends Command {
     public void run(TaskList taskList, Ui ui, Storage storage) throws DiboException {
         ui.sayBye();
