+++ src/main/java/dibo/Command.java
@@ -1,3 +1,4 @@
+package dibo;
 public abstract class Command {
 
     public abstract void run(TaskList taskList, Ui ui, Storage storage) throws DiboException;
