+++ src/main/java/dibo/Deadline.java
@@ -1,3 +1,5 @@
+package dibo;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
