+++ src/main/java/dibo/DeleteCommand.java
@@ -1,11 +1,12 @@
-public class UnmarkCommand extends Command {
+package dibo;
+public class DeleteCommand extends Command {
     private int taskId;
 
-    public UnmarkCommand(int taskId) {
+    public DeleteCommand(int taskId) {
         this.taskId = taskId;
     }
     public void run(TaskList taskList, Ui ui, Storage storage) throws DiboException {
-        ui.showUnmarked(taskList.unmarkTask(taskId));
+        ui.showDeleted(taskList.deleteTask(taskId), taskList.getSize());
         storage.save(taskList);
     }
