+++ src/main/java/dibo/DiboException.java
@@ -1,3 +1,4 @@
+package dibo;
 public class DiboException extends Exception {
     public DiboException(String errorMessage) {
         super(errorMessage);
