+++ src/main/java/dibo/Event.java
@@ -1,3 +1,5 @@
+package dibo;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
