+++ src/main/java/dibo/Storage.java
@@ -1,3 +1,5 @@
+package dibo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
