+++ src/main/java/dibo/Task.java
@@ -1,3 +1,5 @@
+package dibo;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
