+++ src/main/java/dibo/Ui.java
@@ -1,3 +1,5 @@
+package dibo;
+
 import java.util.Scanner;
 public class Ui {
     private Scanner scanner;
